Finished PS/2 Keyboard driver
The PS/2 keyboard driver now passes the byte genetated at each interrupt to the keyboard scancode finite state machine.  The machine uses a number of states to determine what sequence the scancode is a part of, and when a valid sequence is reached, it notes the key code for that given sequence.  Then, that key code is passed to the keyboard event manager which will be used to store the key events in a buffer until requested.

Refactored the key_codes.h to better fit with C standards.

Relevant changes to the make.config file to reflect the new objects to build.