Redid how paging is handled.
The kernel is now assumed to take up the entire first 4MiB of RAM.  This means it is now officially a lower-half kernel, and will remain that way.

The way that paging gets initialized has been redone.  It now takes into account the fact that the page directory (and tables) itself should be mapped to paged memory.

It is now possible for new blocks to be mapped in paged memory.  This means that I can go forward and redo memory allocation to allow for contiguously mapped memoryspace.